reorganize multiselect files
