+++ client/src/locales/en/messages.po
@@ -134,8 +134,8 @@ msgstr "An “eviction filing” is a legal case for eviction commenced by a lan
 #~ msgid "An “eviction filing” is a legal case for eviction commenced by a landlord against a tenant in Housing Court. Such a case can be commenced for nonpayment of rent (most commonly) or for a violation of the lease (such as a nuisance). The eviction filings number only represents cases filed in Housing Court (data from the New York State Office of Court Administration<0>via the Housing Data Coalition</0> in collaboration with the <1>Right to Counsel Coalition</1>) and not evictions carried out by NYC Marshals.<2/>If you or someone you know is facing eviction and want to learn more about your rights, head over to <3>Housing Court Answers for more information</3>.<4/>Due to privacy restrictions on the use of these data, eviction filings cannot be shown for buildings with fewer than 11 units."
 #~ msgstr "An “eviction filing” is a legal case for eviction commenced by a landlord against a tenant in Housing Court. Such a case can be commenced for nonpayment of rent (most commonly) or for a violation of the lease (such as a nuisance). The eviction filings number only represents cases filed in Housing Court (data from the New York State Office of Court Administration<0>via the Housing Data Coalition</0> in collaboration with the <1>Right to Counsel Coalition</1>) and not evictions carried out by NYC Marshals.<2/>If you or someone you know is facing eviction and want to learn more about your rights, head over to <3>Housing Court Answers for more information</3>.<4/>Due to privacy restrictions on the use of these data, eviction filings cannot be shown for buildings with fewer than 11 units."
 
+#: src/components/Multiselect.tsx:455
 #: src/components/PortfolioFilters.tsx:159
-#: src/components/multiselect-dropdown/multiselect/Multiselect.tsx:507
 msgid "Apply"
 msgstr "Apply"
 
@@ -246,7 +246,7 @@ msgstr "Class I"
 msgid "Clear Filters"
 msgstr "Clear Filters"
 
-#: src/components/multiselect-dropdown/multiselect/Multiselect.tsx:488
+#: src/components/Multiselect.tsx:441
 msgid "Clear all"
 msgstr "Clear all"
 
@@ -954,7 +954,7 @@ msgstr "Share this page with your neighbors"
 msgid "Shareholder"
 msgstr "Shareholder"
 
-#: src/components/multiselect-dropdown/multiselect/Multiselect.tsx:396
+#: src/components/Multiselect.tsx:352
 msgid "Show less"
 msgstr "Show less"
